+++ test/runtests.jl
@@ -43,6 +43,7 @@ const SAMPLE_MS = get(ENV, "MEASUREMENTSETS_TEST_MS",
     include("dysco_tests.jl")
     include("empty_undefined_tests.jl")
     include("keywords_tests.jl")
+    include("endian_tests.jl")
     include("container_tests.jl")
     include("taql_query_tests.jl")
     include("taql_command_tests.jl")